fix source ordering

+++ cpp/src/arrow/util/async_generator.h
@@ -1501,7 +1501,7 @@ AsyncGenerator<T> MakeConcatenatedGenerator(AsyncGenerator<AsyncGenerator<T>> so
 template <typename T>
 struct Enumerated {
   T value;
-  int index;
+  int64_t index;
   bool last;
 };
